Move keypair creation function into proper module
